+++ test/findings/2026-02-18/one/SEC_H_04_ExcessiveDisputePayout.t.sol
@@ -62,9 +62,9 @@ contract SEC_H_04_ExcessiveDisputePayout is BaseTest {
         (, uint256[] memory indices) = _claimAndContribute(contributor1, projectId, 1);
         uint256 idx = indices[0];
 
-        _commitAndReveal(validator1, projectId, idx, 3000, uint128(VALIDATOR_STAKE));
-        _commitAndReveal(validator2, projectId, idx, 3000, uint128(VALIDATOR_STAKE));
-        _commitAndReveal(validator3, projectId, idx, 3000, uint128(VALIDATOR_STAKE));
+        _commitAndReveal(validator1, projectId, idx, 3000, VALIDATOR_STAKE);
+        _commitAndReveal(validator2, projectId, idx, 3000, VALIDATOR_STAKE);
+        _commitAndReveal(validator3, projectId, idx, 3000, VALIDATOR_STAKE);
         engine.computeConsensus(projectId, idx);
 
         uint256 escrowAfterConsensus = engine.getProjectEscrow(projectId, address(token));
@@ -138,9 +138,9 @@ contract SEC_H_04_ExcessiveDisputePayout is BaseTest {
         (, uint256[] memory indices) = _claimAndContribute(contribNum == 0 ? contributor1 : contributor2, projectId, 1);
         uint256 idx = indices[0];
 
-        _commitAndReveal(validator1, projectId, idx, 3000, uint128(VALIDATOR_STAKE));
-        _commitAndReveal(validator2, projectId, idx, 3000, uint128(VALIDATOR_STAKE));
-        _commitAndReveal(validator3, projectId, idx, 3000, uint128(VALIDATOR_STAKE));
+        _commitAndReveal(validator1, projectId, idx, 3000, VALIDATOR_STAKE);
+        _commitAndReveal(validator2, projectId, idx, 3000, VALIDATOR_STAKE);
+        _commitAndReveal(validator3, projectId, idx, 3000, VALIDATOR_STAKE);
         engine.computeConsensus(projectId, idx);
 
         vm.prank(challenger);
